ChromeDriver returns cookie expiry dates as floats
We change handling of the Expiry field of returned cookies to handle
both integer and float formats.

The following ChromeDriver commit changed the expiry field from an int
to a double to handle times far in the future.

https://chromium.googlesource.com/chromium/src.git/+/6cdce5554de9e4b686afc4cd9a934491ebae7fcc

The aspirational standard says it should be an int:

https://www.w3.org/TR/webdriver/#cookies

However, the various other clients seem to handle both. For example,
here is Java's implementation:

https://github.com/SeleniumHQ/selenium/blob/ab76d0ff9b38455e199d775e4e9da662eebb3f43/java/client/src/org/openqa/selenium/remote/RemoteWebDriver.java#L753