Improve local testing: run Selenium at test startup
This commit makes local testing more useful. By default, we now start an
Xvfb process in the background, then start a Selenium WebDriver server,
and start executing the tests for both Firefox and Chrome.

The framework to start the background processes is similar to what is
provided in other languages, so they should be reusable downstream for
testing actual web pages.

We also support testing within a Docker container via a command line
flag provided to the `go test` invocation.

Because of the variability of development workstations, we supply a
number of additional command line flags to configure the testing
environment.

More details of specifically how to invoke the tests are in the
README.md.

We add the start of a ChromeOptions struct that was needed to provide an
argument to Chrome running under Docker. That struct will become more
complete in an upcoming commit.

Finally, we add a utility program to download the WebDriver Jars and the
ChromeDriver binary.