Slightly modify copyright assignment implementation
Contributions to this project are licensed through the MIT license in
LICENSE.txt by the individual making the contribution, who retains
the contribution's copyright.

In some cases, individuals will actually be representing their
corporations. Make this case clearer by:

1) Making the copyright statement more correct.
2) Correcting my own entry in AUTHORS.

Also, add a disclaimer that this is not an official Google project,
despite my involvement with the project and employment with the company.