Improve init.go to remove dependency on system-installed Firefox
In order to better control which version of Firefox to test against,
download two versions explicitly: the most recent version supported by
Selenium 2 and a recent nightly. The latter will be used for Selenium 3
testing.

While we're here, also download other files for future use.

We continue to use the Travis-installed Firefox version.