Integrate with Sauce Labs
Sauce Labs provides a browser-based testing environment in the cloud.
This makes it easy to test with many different browsers, browser
versions, operating systems, and operating system versions.

We start with a single simple case of testing the latest Firefox version
on Windows, which is a platform for which we previously had no test
coverage.

To allow us to continue testing using our dummy, in-process web server,
we also implement a type that starts the Sauce Connect Proxy as a
subprocess.

Sauce Labs testing is off by default.

Also, migrate Selenium 2 testing to be performed via Sauce Labs. This
removes some dependencies and ugliness from testing.

Issue #77.
Issue #74.