Expand the example to be runnable, with modification
Go supports two different types of examples: one which is compiled, but
not run, and another which is actually run.

We expand the existing example test to be of the latter type with a
small modification, to better support debugging newcomers to the
package.

The problem with making this example test runnable by default is that is
is not hermetic: it is dependent on files existing in particular
locations and requires network access. The latter can be alieviated by
starting an internal HTTP server, but the binary dependency is
necessarily unavoidable.

Issue #88