Support the `W3C` capability for Chrome
Previously, `AddChrome` would quietly override `W3C` to `true`,
even if you wanted it to be `false`.

Due to a problem with Chrome, the W3C standard, and `elementIsDisplayed`,
it is necessary to be able to disable W3C mode for Chrome so that
we can call `IsDisplayed`.

This behavior appears to have started around Chrome 75.

See:
* https://stackoverflow.com/questions/56111529/cannot-call-non-w3c-standard-command-while-in-w3c-mode-seleniumwebdrivererr
* https://github.com/webdriverio/webdriverio/issues/4073