Use known browser versions
This updates `vendor/init.go` to use known browser versions, by default.
You may still use the latest nightlies by passing `--download_latest`.

The driving force behind these changes was to couple the `chromedriver`
file with its matching `chrome` version.  Starting with Chrome 73,
the `chromedriver` must be matched with the browser version.

I also noticed that because this pulls the browser nightlies, it is
very difficult to reproduce any issues exactly, since each run of
this in GitHub will result in a different browser.  By default, a
version of Chrome and a version of Firefox have been pinned, just like
all of the other binaries that are downloaded.