Move test functions into a dedicated package for re-use
This will allow external WebDriver harnesses the ability to re-use the
existing tests to ensure compatibility.

At some point in the future, we should move the seleniumtest package out
of internal/. The only consumer at the moment does not respect the
internal package convention, so it doesn't matter to them.